Back: -Implementación general de DVH -Recálculo en operaciones de persistencia -(Falta configurar DVV) -Corrección y mejora de bitácoras -(Falta añadir CorrelationID en los errores como parámetro)
Front:
-Diseño e implementación de barra colapsable (todavía no es genérica)
NOTA: No se hizo un"achicador" de íconos sino que se subieron nuevos directamente como imágenes.
-Corección de márgenes de 0px a 3px para dar un sutil aire entre ToolStrips y el borde del formulario (se aplica sobre ToolStrips y no sobre el panel que las contiene).
-Lógica condicional (máquina de estados) para que el atajo Ctrl+B funcione en la nueva implementación de formaa coherente.
-Se toca el DGVCustomForm para que pueda "ceder" la SearchBar removiéndola de sus controles y permitiendo que otro control la use y devuelva.

+++ Domain/Entities/Payments/Payment.cs
@@ -70,7 +70,7 @@ namespace Domain.Entities
 
             return string.Join("|",
                 Id.ToString(),
-                Amount.Value.ToString("F2", culture),
+                ((decimal)Amount.Value).ToString("F2", culture),
                 CreationDate.ToString("yyyyMMddHHmmss", culture),
                 EffectiveDate.ToString("yyyyMMddHHmmss", culture),
                 SaleId.ToString(),
